Tested DD, and it passed.
I need to make documentation for the collision system, as well as fix documentation for the DD, and fix the responsibilities of each class

+++ include/Device.hpp
@@ -53,7 +53,7 @@ public:
 
     /// @brief Updates this device
     /// @param deltaTime 
-    virtual void update(long long deltaTime) = 0;
+    virtual void update(double deltaTime) = 0;
 
     /// @brief Gets the id 
     /// @return a copy of the id
